Refactored and Modernised aspects of the app
- Changed package to com.infinity.foodstandards
- Added Androidx Navigation support
- Using suspend, dispatchers, and scopes for network requests
- Updated consistent use of viewbinding
- Removed activity based interfaces for callbacks
- General code cleanup and readability
