+++ .idea/misc.xml
@@ -1,4 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="app/src/main/res/layout/authorities_fragment.xml" value="0.1641337386018237" />
+        <entry key="app/src/main/res/layout/local_authority_item.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/main_activity.xml" value="0.1641337386018237" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="11" project-jdk-type="JavaSDK" />
 </project>
