+++ app/build.gradle
@@ -7,7 +7,7 @@ apply from: "${gradleDir}/ktlint.gradle"
 android {
     compileSdkVersion 32
     defaultConfig {
-        applicationId "com.aaron.infinity_foodstandards"
+        applicationId "com.infinity.foodstandards"
         minSdkVersion 19
         targetSdkVersion 32
         versionCode 1
@@ -38,6 +38,11 @@ android {
             buildConfigField("Boolean", "MOCK", MOCK)
         }
     }
+    sourceSets {
+        main.java.srcDirs += 'src/main/kotlin'
+        test.java.srcDirs += 'src/test/kotlin'
+        androidTest.java.srcDirs += 'src/androidTest/kotlin'
+    }
     buildFeatures {
         viewBinding true
     }
@@ -56,6 +61,8 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'android.arch.lifecycle:viewmodel:1.1.1'
     implementation 'android.arch.lifecycle:extensions:1.1.1'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.1'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.1'
 
     implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'com.squareup.retrofit2:retrofit:2.8.1'
@@ -65,6 +72,7 @@ dependencies {
     testImplementation 'androidx.arch.core:core-testing:2.1.0'
     testImplementation 'org.mockito:mockito-core:3.7.7'
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.8.1'
+    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.4'
 
     androidTestImplementation 'androidx.test:runner:1.4.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
