--- app/src/main/java/com/aaron/infinity_foodstandards/model/LocalAuthority.kt
@@ -1,20 +1,0 @@
-package com.aaron.infinity_foodstandards.model
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-class LocalAuthority {
-
-    @SerializedName("Name")
-    @Expose
-    lateinit var name: String
-
-    @SerializedName("LocalAuthorityId")
-    @Expose
-    var id: Int = 0
-
-}
-
-class LocalAuthoritiesResponse {
-    lateinit var authorities: List<LocalAuthority>
-}
