--- app/src/main/java/com/aaron/infinity_foodstandards/ui/AuthorityListener.kt
@@ -1,8 +1,0 @@
-package com.aaron.infinity_foodstandards.ui
-
-interface AuthorityListener {
-
-    //Called when a city is clicked
-    fun onAuthorityClick(authorityId: Int)
-
-}
