+++ app/src/main/kotlin/com/infinity/foodstandards/MainActivity.kt
@@ -1,0 +1,13 @@
+package com.infinity.foodstandards
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.main_activity)
+    }
+
+}
