+++ app/src/main/kotlin/com/infinity/foodstandards/model/Establishment.kt
@@ -1,0 +1,18 @@
+package com.infinity.foodstandards.model
+
+import android.os.Parcelable
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Establishment(
+    @SerializedName("RatingKey")
+    @Expose
+    var rating: String
+): Parcelable
+
+class EstablishmentResponse {
+    lateinit var establishments: List<Establishment>
+}
+
