+++ app/src/main/kotlin/com/infinity/foodstandards/model/LocalAuthority.kt
@@ -1,0 +1,20 @@
+package com.infinity.foodstandards.model
+
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+class LocalAuthority {
+
+    @SerializedName("Name")
+    @Expose
+    lateinit var name: String
+
+    @SerializedName("LocalAuthorityId")
+    @Expose
+    var id: Int = 0
+
+}
+
+class LocalAuthoritiesResponse {
+    lateinit var authorities: List<LocalAuthority>
+}
