+++ app/src/main/kotlin/com/infinity/foodstandards/network/FoodStandardsApi.kt
@@ -1,0 +1,14 @@
+package com.infinity.foodstandards.network
+
+import com.infinity.foodstandards.model.LocalAuthoritiesResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Headers
+
+interface FoodStandardsApi {
+
+    @Headers("x-api-version:2")
+    @GET("authorities")
+    suspend fun getAuthorities(): LocalAuthoritiesResponse
+
+}
