+++ app/src/main/kotlin/com/infinity/foodstandards/ui/hygieneRating/HygieneRatingsViewModel.kt
@@ -1,0 +1,17 @@
+package com.infinity.foodstandards.ui.hygieneRating
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.ViewModel
+import com.infinity.foodstandards.model.EstablishmentResponse
+import com.infinity.foodstandards.network.FoodStandardsRepo
+import com.infinity.foodstandards.network.RetrofitService
+
+class HygieneRatingsViewModel : ViewModel() {
+
+    private var foodStandardsRepo: FoodStandardsRepo =
+        FoodStandardsRepo(RetrofitService.createService())
+
+    fun getAllEstablishments(id: Int): LiveData<EstablishmentResponse> {
+        return foodStandardsRepo.getEstablishments(id)
+    }
+}
