+++ app/src/main/res/layout/authorities_fragment.xml
@@ -12,6 +12,8 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:scrollbars="vertical"
+        android:orientation="vertical"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
@@ -20,4 +22,5 @@
         app:layout_constraintVertical_bias="0.0"
         tools:itemCount="50"
         tools:listitem="@layout/local_authority_item" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
