--- app/src/test/java/com/aaron/infinity_foodstandards/ui/hygieneRating/MockitoUtils.kt
@@ -1,5 +1,0 @@
-package com.aaron.infinity_foodstandards.ui.hygieneRating
-
-import org.mockito.Mockito
-
-inline fun <reified T> mock(): T = Mockito.mock(T::class.java)
