+++ app/src/test/kotlin/com/infinity/foodstandards/ui/hygieneRating/MockitoUtils.kt
@@ -1,0 +1,5 @@
+package com.infinity.foodstandards.ui.hygieneRating
+
+import org.mockito.Mockito
+
+inline fun <reified T> mock(): T = Mockito.mock(T::class.java)
